// frontend/src/types/product.ts
export type productType = {
    product: {
        _id: string;
        name: string;
        image: string;
        description: string;
        brand: string;
        category: string;
        price: number;
        countInStock: number;
        rating: number;
        numReviews: number;
    }
}

export type ProductType = {
    _id: string;
    name: string;
    image: string;
    description: string;
    brand: string;
    category: string;
    price: number;
    countInStock: number;
    rating: number;
    numReviews: number;
}